[BE] refactor: 파일 경로 및 절대 경로 수정
