+++ backend/src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { ConfigModule, ConfigService } from '@nestjs/config'
-import { AppController } from './app.controller'
-import { AppService } from './app.service'
-import { Event } from './entities/event.entity'
-import { EventModule } from './modules/event.module'
+import { Event } from '@entities/event/event.entity'
+import { EventModule } from '@modules/event/event.module'
+import { AppController } from '@src/app.controller'
+import { AppService } from '@src/app.service'
 
 @Module({
   imports: [
