+++ backend/src/controllers/ticket/ticket.controller.ts
@@ -1,8 +1,8 @@
 import { Controller, Get, Post, Body, Param } from '@nestjs/common'
 import { ApiCreatedResponse, ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger'
-import { CreateTicketDto } from '@dto/ticket.dto'
-import { Ticket } from '@entities/ticket.entity'
-import { TicketService } from '@services/ticket.service'
+import { CreateTicketDto } from '@src/dto/ticket/ticket.dto'
+import { Ticket } from '@src/entities/ticket/ticket.entity'
+import { TicketService } from '@services/ticket/ticket.service'
 
 @Controller('tickets')
 @ApiTags('tickets')
