+++ backend/src/entities/event/event.entity.ts
