+++ backend/src/main.ts
@@ -1,7 +1,7 @@
 import { NestFactory } from '@nestjs/core'
-import { AppModule } from './app.module'
-import { setupSwagger } from './utils/swagger'
-import { corsConfig } from './config/cors.config'
+import { AppModule } from '@src/app.module'
+import { setupSwagger } from '@utils/swagger'
+import { corsConfig } from '@config/cors.config'
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule)
