+++ backend/src/modules/event/event.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
-import { EventController } from '../controllers/event.controller'
-import { EventService } from '../services/event.service'
-import { Event } from '../entities/event.entity'
+import { EventController } from '@src/controllers/event/event.controller'
+import { EventService } from '@services/event/event.service'
+import { Event } from '@src/entities/event/event.entity'
 
 @Module({
   imports: [TypeOrmModule.forFeature([Event])],
