+++ backend/src/modules/ticket/ticket.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
-import { Ticket } from '@entities/ticket.entity'
-import { TicketController } from '@controllers/ticket.controller'
-import { TicketService } from '@services/ticket.service'
+import { Ticket } from '@entities/ticket/ticket.entity'
+import { TicketController } from '@controllers/ticket/ticket.controller'
+import { TicketService } from '@services/ticket/ticket.service'
 
 @Module({
   imports: [TypeOrmModule.forFeature([Ticket])],
